requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
browserforge==1.2.3
camoufox==0.4.11
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
frozenlist==1.7.0
geoip2==5.1.0
greenlet==3.2.3
idna==3.10
language-tags==1.2.0
lxml==6.0.0
maxminddb==2.7.0
multidict==6.6.3
numpy==2.3.1
orjson==3.11.0
platformdirs==4.3.8
playwright==1.53.0
propcache==0.3.2
pyee==13.0.0
PySocks==1.7.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
screeninfo==0.8.1
sentry-sdk==2.34.1
soupsieve==2.7
tqdm==4.67.1
typing_extensions==4.14.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
yarl==1.20.1
